pyproject.toml
[tool.poetry]
name = "capstone"
version = "0.1.0"
description = ""
authors = ["hirokioda <hiroki.odd30302@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
numpy = "^1.25.2"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
isort = "^5.12.0"
black = {extras = ["jupyter"], version = "^23.7.0"}
pyproject-flake8 = "^6.0.0.post1"
pytest = "^7.4.0"
jupyterlab = "^4.0.5"
mypy = "^1.5.0"
mglearn = "^0.2.0"
graphviz = "^0.20.1"
openpyxl = "^3.1.2"
bs4 = "^0.0.1"
lxml = "^4.9.3"
html-parser = "^0.2"
html5lib = "^1.1"
pandas-datareader = "^0.10.0"
statsmodels = "^0.14.0"
spotipy = "^2.23.0"
python-dotenv = "^1.0.0"
networkx = "^3.3"
community = "^1.0.0b1"
tqdm = "^4.66.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
profile = "black"
use_parentheses = true

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.flake8]
max-line-length = 120
ignore = "E203,"

[tool.pytest.ini_options]
testpaths = ["tests",]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
check_untyped_defs = true
show_column_numbers = true
warn_unused_configs = true
warn_unused_ignores = true
namespace_packages = true

[tool.poetry.scripts]
lint = "flake8 path/to/your/code/"
mypy = "mypy path/to/your/code/"
format = "black path/to/your/code/ && isort path/to/your/code/"
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
